first attempt at random number generator

+++ waste/waste_serial.c
@@ -30,7 +30,7 @@ int main(int argc, char *argv[]) {
     sscanf(argv[1], "%i", &ntarget); 
 
     // dynamically allocate arrays
-	xa = (int *)malloc(sizeof(float)*ntarget); 
+	xa = (float *)malloc(sizeof(float)*ntarget); 
 
 	// Equatorial coordinates of accepted events
 	memset(xa, 0, sizeof(int)*ntarget); // initializes array to zeroes
